@@ -1,15 +1,3 @@
-"""
-General approach here is:
--> Load timestamp mapping from Basler data (raw_videos/timestamp_mapping.json)
-    -> use "starting_mapping" to figure out camera_timestamp to utc mapping for each camera
-    -> verify "starting_mapping" with "ending_mapping"
--> Load timestamp mapping from Pupil data (pupil_voutput/info.player.json)
-    -> interested in "start_time_synced_s" (pupil timestamp format) and "start_time_system_s" (utc)
--> Get map from pupil timestamp to each Basler camera timestamp
--> Trim Pupil data to match Basler data
-    -> for now, using synched Basler data (synchronized_videos/timestamps.npy) - may need to take a mean of Basler timestamps? or synch to 1 camera?
-"""
-
 import json
 from pathlib import Path
 from typing import Dict, List, Tuple
@@ -308,13 +296,13 @@ class PupilSynchronize:
         print(f"ending offsets in frames: {ending_offsets_in_frames}")
         return ending_offsets_in_frames
 
-    def save_corrected_timestamps(self):
-        if self.corrected_timestamps is None:
+    def save_synchronized_timestamps(self):
+        if self.synchronized_timestamps is None:
             raise ValueError(
-                "corrected_timestamps is None, this method should only be called from synchronize(), it should not be called directly")
-        for cam_name, timestamps in self.corrected_timestamps.items():
+                "synchronized_timestamps is None, this method should only be called from synchronize(), it should not be called directly")
+        for cam_name, timestamps in self.synchronized_timestamps.items():
             print(f"cam {cam_name} timestamps shape: {timestamps.shape}")
-            np.save(f"{self.output_path}/cam_{cam_name}_corrected_timestamps.npy", timestamps)
+            np.save(f"{self.output_path}/cam_{cam_name}_synchronized_timestamps.npy", timestamps)
 
     def save_metadata(self):
         with open(f"{self.output_path}/metadata.json", "w") as f:
@@ -392,7 +380,7 @@ class PupilSynchronize:
         dropped_frames = 0
         skipped_frames = 0
         early_frames = 0
-        corrected_timestamps: List[int | None] = []
+        synchronized_timestamps: List[int | None] = []
         previous_frame = np.zeros((framesize[1], framesize[0], 3))
 
         while True:
@@ -421,10 +409,9 @@ class PupilSynchronize:
             # if we make it past the if/elif, the current timestamp is between the start and end times
             if current_timestamp > (reference_timestamp + (0.5 * median_duration)):
                 # current frame is too late, don't read it and fill in a dummy frame instead
-                # frame = cv2.drawMarker(previous_frame, (20, 20), (0, 0, 255), cv2.MARKER_STAR, 30, 1)
-                frame = previous_frame
+                frame = cv2.drawMarker(previous_frame, (20, 20), (0, 0, 255), cv2.MARKER_STAR, 30, 1)
                 video_writer_object.write(frame)
-                corrected_timestamps.append(None)
+                synchronized_timestamps.append(None)
                 written_frames += 1
                 dropped_frames += 1
             elif current_timestamp < (reference_timestamp - (0.5 * median_duration)):
@@ -447,7 +434,7 @@ class PupilSynchronize:
                 previous_frame = frame
 
                 video_writer_object.write(frame)
-                corrected_timestamps.append(current_timestamp)
+                synchronized_timestamps.append(current_timestamp)
 
                 written_frames += 1
                 current_frame += 1
@@ -457,7 +444,7 @@ class PupilSynchronize:
               f"\tearly frames: {early_frames}\n"
               f"\tskipped frames: {skipped_frames}")
 
-        self.corrected_timestamps[camera_name] = np.array(corrected_timestamps)
+        self.synchronized_timestamps[camera_name] = np.array(synchronized_timestamps)
 
         cap.release()
         video_writer_object.release()
@@ -484,18 +471,18 @@ class PupilSynchronize:
         )
 
     def verify_framecounts(self):
-        for cam_name in self.corrected_timestamps.keys():
+        for cam_name in self.synchronized_timestamps.keys():
             if cam_name.startswith("eye"):
                 cap = cv2.VideoCapture(str(self.output_path / f"{cam_name}.mp4")) # THIS PATH IS WRONG
             else:
                 cap = cv2.VideoCapture(str(self.output_path / f"{self.index_to_serial_number[cam_name]}.mp4"))
             frame_count = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-            if frame_count != self.corrected_timestamps[cam_name].shape[0]:
+            if frame_count != self.synchronized_timestamps[cam_name].shape[0]:
                 print(
-                    f"frame count mismatch for cam {cam_name}: video: {frame_count} vs timestamps: {self.corrected_timestamps[cam_name].shape[0]}")
+                    f"frame count mismatch for cam {cam_name}: video: {frame_count} vs timestamps: {self.synchronized_timestamps[cam_name].shape[0]}")
             else:
                 print(
-                    f"frame count match for cam {cam_name}: video: {frame_count} vs timestamps: {self.corrected_timestamps[cam_name].shape[0]}")
+                    f"frame count match for cam {cam_name}: video: {frame_count} vs timestamps: {self.synchronized_timestamps[cam_name].shape[0]}")
 
             cap.release()
 
@@ -513,14 +500,14 @@ class PupilSynchronize:
             "ending_offsets_frames": ending_offsets_frames,
         }
 
-        self.corrected_timestamps = {
+        self.synchronized_timestamps = {
             cam_name: self.synched_basler_timestamps_utc[i, :][
                       starting_offsets_frames[cam_name]: ending_offsets_frames[cam_name] + 1
                       ] for i, cam_name in enumerate(self.basler_camera_names)
         }
 
         self.trim_videos(starting_offsets_frames=starting_offsets_frames, ending_offsets_frames=ending_offsets_frames)
-        self.save_corrected_timestamps()
+        self.save_synchronized_timestamps()
         self.verify_framecounts()
 
         self.plot_raw_timestamps(
@@ -528,7 +515,7 @@ class PupilSynchronize:
             ending_offsets=ending_offsets_frames
         )
 
-        self.plot_corrected_timestamps()
+        self.plot_synchronized_timestamps()
 
     def plot_raw_timestamps(
             self,
@@ -574,7 +561,7 @@ class PupilSynchronize:
 
         plt.show()
 
-    def plot_corrected_timestamps(self):
+    def plot_synchronized_timestamps(self):
         """plot some diagnostics to assess quality of camera sync"""
         # TODO: swap time and frame number, so x axis shows synching
         # opportunistic load of matplotlib to avoid startup time costs
@@ -586,12 +573,12 @@ class PupilSynchronize:
         fig.suptitle(f"Timestamps")
 
         ax1 = plt.subplot(
-            title="(Corrected) Camera Frame Timestamp vs Frame#\n(Lines should have same slope)",
+            title="(Synchronized) Camera Frame Timestamp vs Frame#\n(Lines should have same slope)",
             xlabel="Frame#",
             ylabel="Timestamp (ns)",
         )
 
-        for name, timestamps in self.corrected_timestamps.items():
+        for name, timestamps in self.synchronized_timestamps.items():
             ax1.plot(timestamps, label=name)
 
         ax1.legend()
@@ -603,14 +590,10 @@ class PupilSynchronize:
 
 
 if __name__ == "__main__":
-    # folder_path = Path("/Users/philipqueen/Basler_pupil_synch_test")
     folder_path = Path(
         "/Users/philipqueen/ferret_0776_P35_EO5/"
     )
     pupil_synchronize = PupilSynchronize(folder_path)
-
-    # print(pupil_synchronize.basler_timestamp_mapping)
-    # print(pupil_synchronize.pupil_timestamp_mapping)
 
     utc_timestamp_per_camera = pupil_synchronize.get_utc_timestamp_per_camera()
     utc_start_time_pupil = pupil_synchronize.pupil_start_time_utc
@@ -636,19 +619,10 @@ if __name__ == "__main__":
         f"basler start times per camera: {pupil_synchronize.basler_timestamp_mapping['starting_mapping']['camera_timestamps']}"
     )
 
-    # print("basler timestamps (in s since start):")
-    # print(f"{np.min(pupil_synchronize.synched_basler_timestamps) / 1e9}")
-    # print(f"{np.max(pupil_synchronize.synched_basler_timestamps) / 1e9}")
-    # print(f"{np.mean(pupil_synchronize.synched_basler_timestamps) / 1e9}")
-
     print(
         f"pupil timestamps shapes - eye0: {pupil_synchronize.pupil_eye0_timestamps_utc.shape} eye1: {pupil_synchronize.pupil_eye1_timestamps_utc.shape}"
     )
     print(f"pupil timestamps (eye0): {pupil_synchronize.pupil_eye0_timestamps_utc}")
     print(f"pupil timestamps (eye1): {pupil_synchronize.pupil_eye1_timestamps_utc}")
-
-    pupil_synchronize.get_pupil_fps()
-    pupil_synchronize.get_pupil_median_fps()
-    # pupil_synchronize.find_missing_pupil_frames()
 
     pupil_synchronize.synchronize()
